requirements.txt
asgiref
autopep8
pycodestyle
pytz
sqlparse
Unipath
dj-database-url
python-decouple
gunicorn
whitenoise
django-heroku
psycopg2-binary==2.8.4
Django==2.1.15
django-templated-mail==1.1.1
djangorestframework==3.11.0
djoser==2.0.3
ipython
django-allauth==0.36.0
django-crispy-forms
schedule